[csa] Introduce CallFunction
... and use it in IC dispatchers for calling getters and setters which
are known to be JSFunctions.
Use ConvertReceiverMode::kNotNullOrUndefined when applicable.

In case a non-JSFunction callable object is installed as getter or
setter slow stub is used. In order to test this functionality
the CL adds d8.dom.Document_all callable Api object which gets/sets
`globalThis.document_all_property` depending on number of arguments.

Change-Id: Id03b00e3dce848f1e53763b0293860bac1d48910