Remove handled_hint, do all catch prediction in isolate.cc
For async functions/generators, we now check the handler table
at the generator location when walking the promise tree during
debugging. This allows us to centralize all catch prediction code
in isolate.cc, deleting it from lots of other locations where it
is on a more critical path, and remove the handled_hint flag from
promises.

Bug: None
Change-Id: Icfada30c9326dd75195965e11d771bf905ff7e20