[interpreter] Keep for-in's each variable hint in the whole loop
The BytecodeRegisterOptimizer can set a variable hint to a Register
that holds this variable. However, the hint can be flushed by a
branch.

This CL makes the hint of for-in's each variable to survive the flush
in the loop body, and flushes the hint when leaving the loop. We can
do it because we only use the hint to generate enumerated keyed access
bytecode, and it would be fine if the hint doesn't match the actual
variable in the register, as the bytecode handler will check this
parameter anyway.

Change-Id: I2ea7cc19c65c6128b2e599ff8a0f1d4ef5ac0359