[sandbox] Introduce Code::parameter_count
This field is currently only available for optimized code, in which case
it contains the parameter count of the compiled function, similar to
BytecodeArray::parameter_count. For now, this (trusted) parameter count
is only used during deoptimization where we were previously using the
(untrusted) parameter count stored on the SharedFunctionInfo, which
could lead to various issues if the value was modified by an attacker.

Bug: chromium:335544065
Change-Id: If46be3147ac2acc60d1afdbafe7cdc650cfadd09