[interpreter] Introduce VariableHint in BytecodeRegisterOptimizer
The BytecodeRegisterOptimizer can set a variable to a RegisterInfo,
this means that the register holds this variable. However, the hint
can be flushed by branch.

This CL makes the hint to survive across branches by setting "maybe
variable" to RegisterInfo. This means that the variable is ever or
conditionally set to the register.

We use it for generating enumerated keyed access bytecode. We can do
it because the bytecode handler will check this parameter anyway, so
would be fine if the hint doesn't match the actual variable in the
register.

Bug: v8:14245
Change-Id: I2ea7cc19c65c6128b2e599ff8a0f1d4ef5ac0359