[wasm][turboshaft] Emit tail call for invalid functions
When we try to inline a tail call and we encounter an invalid inlinee,
we should emit a tail call (over a regular call) as a replacement.
This bug can manifest with lazy validation, when a tail call is never
called and thus the callee is not validated by liftoff before tier-up.

Bug: chromium:334687959
Change-Id: Idf299d81d1da6e590558129690c94c20f3e08666