Implement v8::Date::Parse and v8::Date::ToUTCString
This patch adds two additional methods to the v8::Date C++ API. The
first is v8::Date::Parse, which is a static method that takes a string,
parses it, and returns the v8::Date instance. This is equivalent to
using `new Date(str)` in JavaScript. The second is

v8::Date::ToUTCString(), which returns the UTC string representation
of the date similar to the existing ToISOString() method.
Change-Id: I0f4ab7d5f01ca549a6a1a23ec9bee10a6880d64e