[wasm][turbofan] Fix wrong types for memory.size
By always picking the first memory to decide on the integer type of
memory.size, we might either wrongly explicitly truncate it and
implicitly extend it or wrongly explicitly extend it to then
implicitly truncate it.

In either case, Turbofan does not seem care about these type
differences and the value can never be larger than int32 max, so this
is just an inconsistency.

Bug: 42204191
Change-Id: Ib04a8dccb0f0faa63930211136f0b5acf3ed89a2