[flags] let FLAG_VALUE_SCOPE support multiple flag types
Exposes the underlying flag type as FlagValue::type, and uses this to
allow FLAG_VALUE_SCOPE() to provide a define a FlagScope with the
underlying type of the FlagValue, and which requires a matching
parameter type as the second macro argument. This expandds the
functionality of the test helper macro, with some added type safety
benefits.

Change-Id: I84b79b059c1dbb29238d5579cc3706e70adae8b8