[tools] Fix gen-keyword-gen-h.py and improve error message
Running the script on main gives the following error:

"Didn't get exactly %d replacement(s) for pattern: %s"

because of a upper case only check where it should accept lowercase
too.

Additionally, changed the asso_values array length to 129 as that is
the length of the output produced by gperf.

Drive-by: Improve error message to display the expected number of
replacements.

Change-Id: I9df49190b8209869bbe9ba2fa8ed2e9f861ff6e0