[wasm] Refactor call feedback entry layout
Small, semantics-preserving refactorings:
1. For `call_ref` feedback, determine whether an entry is
uninitialized by a Smi(0) in the second slot, not the first slot.
This prepares for `call_indirect` feedback, where we will use a Smi
in the first slot for (the lower bits of) the call target address, so
Smi(0) could be a genuine monomorphic feedback entry and no longer
just a sentinel for uninitialized feedback entries.
2. Make the second slot always contain a Smi, also in the mono- and
polymorphic `call_ref` feedback cases.
3. Make the second slot contain the call count for both direct calls
and monomorphic `call_ref` feedback (was only the latter before).

Drive-by: Also add a bunch more `DCHECK`s when processing the feedback.

Bug: 335082212
Change-Id: I6e957c916ca667c583f16a297de9c5844954f1a3