[sandbox] Forbid resetting external pointer fields
Until now, we allowed resetting (lazily-initialized) external pointer
fields, in which case we would set the handle to the null handle. This
was mostly used as a small optimization when detaching array buffers.
However, it required some additional complexity in the external pointer
table (EPT), namely during compaction. With the move to managed external
resources, this becomes even more of a problem as the following could
happen (using the example of ArrayBuffers and their extension):

1. We run EPT compaction and mark the entry of an ArrayBufferExtension
   for evacuation (i.e. create an evacuation entry for it)
2. We detach the ArrayBuffer, setting the ArrayBuffer's external pointer
   to the null handle
3. We sweep the EPT, but now skip the entry of the ArrayBufferExtension
4. If the segment that the EPT entry was in is now completely empty, we
   will deallocate the segment
5. When the ArrayBufferExtension is subsequently destroyed, it tries to
   zap it's EPT entry, which now no longer exists and we crash

Dealing with this is somewhat challenging, and so instead of adding more
complexity to the EPT, with this CL we now simply forbid resetting
external pointer fields. As such, once an external pointer field is
initialized and has a corresponding entry in the EPT, it must keep that
entry until it is destroyed.

Bug: chromium:327550517
Change-Id: Ib74cd174d7aa2a97ba0ae14afd56a1ae9999c753