[sandbox] Use Code::parameter_count in more places in the deoptimizer
As the SharedFunctionInfo::internal_formal_parameter_count() is
deprecated, this CL switches a few more places in the deoptimizer to use
the parameter count from the Code object instead. It also now adds a
code comment to SharedFunctionInfo::internal_formal_parameter_count() to
note that the field is deprecated and should no longer be used.

Bug: chromium:40931165
Change-Id: Ic5cdc1370d2b8d673289b9b110cf1aa3539efa9c