[heap] Enable background threads in marking steps
This CL allows to re-enable the usage of background threads at
incremental marking steps and at the beginning of a GC. Before this
CL the decision whether to use background threads or not was done only
at incremental marking start and would only change with the next GC.
This CL is intended to fix/mitigate regressions around tab
switching. A full GC cycle can take quite some time, so the idea here
is that the GC still operates without background threads while the
tab already runs in foreground.

Bug: 40278463
Change-Id: I2a579d58394e823f06fbeb360ae38858a92727f1