[interpreter/maglev] Skip unnecessary SetPendingMessage around finally
Previously try/finally would always save/restore current pending
messages independent of whether we had caught an exception. Swaps the
single SetPendingMessage for the more explicit ClearPendingMessage,
which returns the current message and sets the current message to "the
hole"; and a RestorePendingMessage, which takes in a previous saved
message.

On paths that don't rethrow, the message to be restored is explicitly
set to "the hole" without calling into ClearPendingMessage first. The
try/finally's catch block is always started with a ReThrow command so
will always clear; but e.g., return in try will jump to finally while
skipping catch, and passes "the hole" as the message to be restored.

In Maglev catch blocks that weren't ever executed are eliminated. If we
only ever see paths that pass "the hole" to RestorePendingMessage, we
can eliminate the call to RestorePendingMessage.

Change-Id: I029659262ccff49d319cc8c60667f467d69789c5