[wasm][code flushing] Add histogram for flushed code size (in bytes)
All Liftoff code is flushed on emergency GCs for allocations and on
memory pressure. This CL introduces a histogram that collects the code
size that is being flushed at these events.

About the histogram:
- we want to see the cases when there is no Wasm code being flushed
  (hence minimum size 0)
- based on V8.WasmModuleSizeBytes.wasm the module can go up around
  almost half GB (hence the max size)

Other histograms that we can look at to see potential negative effects
of flushing are the V8.WasmNumLazyCompilations*Sec histograms.

Bug: 339120678
Change-Id: I176c7c618100a8024884a844cabd1953205ae5e1