Merged: [sandbox] Check array initial map was not corrupted
DependOnInitialMapInstanceSizePrediction uses the default initial
map for the array_function to calculate the slack size.

Later we update the initial map depending on the elements kind.
We assume that the slack calculated by this updated initial map
is the same.

So in the CL we add a check in CapturedObject::set to be sure
we don't overflow the CapturedObject.

Drive-by: We refactor ReduceArrayConstructor to make the
initial map logic more explicit.

Fixed: 339517309
(cherry picked from commit 6615a7a9d0b16a8e6d7611431d85cfefe0e6acbe)

Change-Id: I3d21279d974b80eb2557dbcc129e79cbb1526ec9