[wasm][turboshaft][fix] Return member to avoid signalling NaN flipping to quiet NaN
To get the float32 or float64 values from the Storage, we call the
corresponding getter methods that return a float or a double. This
causes issues if the value is a signaling NaN which gets converted to a
quiet NaN on ia32.
This CL is a quick fix to return the member itself instead of calling
the getters.

Fixed: 339704607
Change-Id: I811be14e6553a2f80fcc12754fa166029375f92e