[baseline|maglev] Collect branch feedback and use it in Maglev
This CL adds feedback collection to the following bytecodes:
JumpIfTrue
JumpIfFalse
JumpIfTrueConstant
JumpIfFalseConstant

We collect the branch feedback in baseline (not in interpreter).

If Maglev detects that a branch was never taken, it generates
a check (that the condition doesn't result in taking the branch), and
doesn't generate code for the never-taken block.

Follow-ups:
- Add feedback-collecting to other Jump bytecodes
- Use the feedback in other ways, e.g., reorder basic blocks.

Bug: 340100647
Change-Id: I04db082c457dcda409a3839700a6f9ba3e82c862