[heap][handles] Revise parking invariants for direct handles
Until now, when parking a background thread, a DCHECK would fail if
the thread's stack contained direct handles, practically allowing
only limited use of direct handles in background threads. With
https://crrev.com/c/5331543, the stacks of parked background threads
are correctly scanned (in builds with conservative stack scanning
enabled) and therefore this restriction can now be eliminated.

This CL:

- allows unrestricted use of direct handles in background threads;
- revises the methods and scopes for parking threads, to follow the
  new invariants described in the design doc (linked from the
  tracking bug b/41480448); and
- updates tests that used the previous logic.

Change-Id: Ia01343d5685cfe4537202e8940bf0cd7d614d8fc

Bug: b/41480448
Bug: b/42203211
Change-Id: Ia01343d5685cfe4537202e8940bf0cd7d614d8fc