[heap] Rename methods for parking threads
Since https://crrev.com/c/5563331, parking background threads in the
context of conservative stack scanning is safe in general, and not
only for blocking (see the design doc, linked from the tracking bug
bug b/41480448). This CL renames the methods used for this purpose,
from "Block*WhileParked*" to "Execute*WhileParked*".

Bug: b/41480448
Change-Id: Icdf0092463956d61bbc8e63e1665348efc40f03f