[turboshaft] Fix VariableReducer bug with variables introduced in loops
The VariableReducer inserts PendingLoopPhis for non-loop-invariant
Variables that are valid before loops, and, when emitting the
backedge, it fixes the PendingLoopPhis that it had introduced before.

However, when a Variable is created inside of a loop, the Variable
reducer was trying to fix its PendingLoopPhi without realizing that it
has in fact none.

Bug: 42204525, 42202729
Change-Id: I5e26503bba797120eb33b547baee2df6fbcea640