[sandbox] Check functionData type in generic JSToWasmWrapper
When accessing the SharedFunctionInfo::Data() field, the result must be
checked to have the expected type. This is because even though the field
contains a trusted pointer, that pointer is accessed using the
kUnknownIndirectPointerTag, for which the documentation states "NOTE:
this tag effectively disables the built-in type-checking mechanism. As
such, in virtually all cases the caller needs to perform runtime-type
checks (i.e. IsXyzObject(obj))` afterwards". This check was so far
missing in the generic JSToWasm wrapper.

To implement this, this CL also adds a `sbxcheck` to torque which, in
line with the SBXCHECK macro, becomes a dcheck in non-sandbox builds.
Further, it tweaks the AbortCSADcheck runtime function to take the
global AbortMode into account, in-line with V8_Fatal.

Bug: 343749048
Change-Id: I2cbcee432c73d025d8bfaa7cc799a3355f8c4cd6