[turboshaft] Prepare support for lazy deopt on first throw
Since https://crrev.com/c/5516582, Maglev lazy deopts the first time a
throw to a so-far unused catch block happens.

This makes graph smaller, and allows more optimizations (because
effects in the catch block then don't pollute the
analysis/optimizations, since the catch block is not even emitted).

This CL adds support for this mechanism in Turboshaft, but only
enables it for the Maglev->Turboshaft pipeline.

Bug: 42204525
Change-Id: I625eea718e56d7b13d5705f87ab5cff4fd97625e