[gc] Change storage of callee_pc in the state of frame iteration
So far the State struct in frames.cc stored the callee_pc_address, and
when the callee_pc was read during frame iteration, it was read just in
time from the callee_pc_address. This approach causes problems on arm64
because there the PC is authenticated, at least most of the time.
However, fast API calls store their PC (the return address) on the
isolate and not on the stack, and the PC on the isolate is not
authenticated. Therefore, during frame iteration, a PC loaded from the
isolate has to be handled differently than a PC loaded from the stack.
At the code locations where the callee_pc is needed it is hard to tell
if the callee_pc_address is on the isolate or on the stack. However,
where the callee_pc_address gets initialized, it is easy. Since the
callee_pc_address is not needed anyways, with this CL the
callee_pc_address does not get stored anymore in the State, but instead
the callee_pc directly.

Bug: 341382353
Change-Id: Ia89b8c03657a8bb101adf57a331c56d957a82345