[maglev] Fix clobbered register in StringCodePointAt
The {string} input and the result could register, which make the
StringLength calculation buggy.

If we are calculating the CodePointAt of a two byte string, we need
a second temporary (or we would need to support adding constrains
like input cannot alias with result in the register allocator).

Drive-by: Make BuiltinStringPrototypeCharCodeOrCodePointAt
arch-agnostic.

Fixed: 345547970
Change-Id: Id42fc7f244a0307699ee39707e7a5b892a4cedf2