[parser] Remove AST/Parser support for JS Runtime calls
JS runtime calls are a vestige of the days when V8 had part
of its runtime implemented in JavaScript. This cleanup CL
removes parser/AST support JS runtime calls and removes the
special-casing for those slots in native contexts.

There are still two uses of CallJSRuntime in the interpreter, but
they hardcode their context indices. This CL updates the unittests
to hardcode the same indices for printing bytecode expectations.

Change-Id: I05756bf6575bcb93b0e9a89407a7239fb0f5b30d