[fastapi] Store the isolate in the options
The isolate is needed to throw an exception, e.g. when a type check
failed. So far the isolate was retrieved from the receiver, but it is
not always possible to retrieve the isolate from the receiver, e.g. when
the receiver is a Smi and not a HeapObject.

With this CL the isolate is passed as a field in the
`FastApiCallbackOptions` struct.

Bug: 41492790
Change-Id: Ie60aee616b3440a92038e4cd36a17316a37bc585