Reland "[turboshaft][csa] Enable Turboshaft builtin pipeline with profiling"
This is a reland of commit d475deeb5740727540253a40d3e2d06bc2fa65fc

No change in this reland. A moderate increase in binary size is
expected due to the additional Turboshaft builtin logic. This will be
temporary as this regression should recover somewhat when we can
delete the corrensponding Turbofan version.

Original change's description:
> [turboshaft][csa] Enable Turboshaft builtin pipeline with profiling
>
> This CL adds profiling support for Turboshaft' builtin pipeline and
> flips the flag to build builtins with this new pipeline.
>
> Change-Id: I921135075f703f6604507ab6a5dc206678bb9669
> Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5596269
> Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
> Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#94398}

Change-Id: I6877c646b76c7544a0ef4957a7d4595c2d1280a0