[heap] Prepare ArrayBufferSweeper for sticky markbits (variant2)
When detaching an ArrayBufferExternsion, we check the age of the
JSArrayBuffer to know which list of ArrayBuffers should contains the
extension. With sticky markbits, if detaching an extension during full
GC marking, the age check cannot be relied on. All objects are unmarked
and treated as old during marking, yet the extension may reside in the
young list.

Instead we need to find the list containing the extension without asking
the JSArrayBuffer. This CL achieves that by storing the age in the
extension, updating it when promoting the extension, and querying it
when detaching.

Bug: 333906585
Change-Id: I2afadd90127e109dd0a61cd61d4cd75e09cea2b4