[builtins] Temporarily switch back to Turbofan builtins
We experience build failures of chrome/ci/orderfile-clankium on arm32.
One potential candidate for the issue that landed in the regression
range is https://chromium-review.googlesource.com/c/v8/v8/+/5625903.
If this is indeed the culprit, it is most likely that it is caused by
switching to Turboshaft for the builtin compiler (--turboshaft-csa), so
this CL is temporarily flipping this flag back. If this doesn't resolve
the issue, we need to consider reverting the entire CL (or looking
elsewhere).

Bug: chromium:347223361
Change-Id: I4ccd6bf63e3e6341569c8504446918da77167566