[deoptimizer][wasm] Allocate feedback vector during deopt if missing
In cases of multiple instantiations of the same module it is possible
that a function is deopted for which the feedback vector has not been
allocated yet.

The deoptimizer needs to allocate it, however during deoptimization a
heap allocation is not possible as the stack is in a non-gc-safe state.

Therefore, the deoptimizer continues at an additional piece of liftoff
code that calls a builtin (with safetpoint information) to allocate the
feedback vector.

Bug: 42204618
Change-Id: I6c67fb3b05061b2e041978dc541674c0191c7fb3