[regexp] Improvements to Unicode case independent.
Case-desugar the regexp later for better performance.

We used to expand into case variants at a very early
stage, in the parser. This has two disadvantages.

Firstly, it means we do all the case work even when
just checking the syntax of a regexp, which happens
multiple places in order to give early errors.

Secondly, it disables some of the optimizations that
Irregexp can do, because now instead of literal strings
we have sequences of character classes, which are not
treated in the same way.

For character classes we have to desugar early to get
the specified semantics, but for literal texts outside
of character classes we can mostly go back to the old
behaviour that is used without the /u and /v flags.
The exception is for surrogate pairs (code points
above 0x10000) where we still desugar early.  Luckily
there are no alpha letters below 0x10000 that have
case equivalents above 0x10000.

Bug: 40261789
Change-Id: I16dca4830c3145f436f537c1db1a0de6def9045c