[fuzzer][wasm] Add new fuzzer for wasm deopts
The fuzzer aims to provide coverage for wasm deopts by emitting
patterns that may lead to deopts.
It emits:
- a global that stores the call target for the call_ref instruction
  which will trigger the deopt
- multiple call targets with the same signature
- (optional) inlinees that potentially get inlined into the optimized
  function and call the other inlinees to the innermost one which
  emits the deopt point (call_ref)
- the main function that calls the inlinees or, if no inlinees are
  emitted, emits the deopt point itself

The fuzzer performs a reference run first without any tier up
iterating over the different call targets, storing them in the global
and calling the main function. All results are collected. Any
long-running loops or non-determinism in this run stops the reference
run and limits until which call target the actual run can be
performed.

The actual run instantiates a new module that uses the same call
target sequence but triggers tier up after the first call target
potentially leading to deopts in subsequent runs.

As the fuzzer might emit a bunch of trapping instructions or control
flow that results in the deopt point becoming unreachable, it is not
guaranteed that any given fuzzer run actually performs a deopt.

Bug: 42204618
Change-Id: I53bf0423865116d748fa54102ce60ab9a31afad3