[maglev] Fix skipped smi check due to phi hoisting
Given we have a phi (29) assumed to be smi at graph building time, we
must not untag it's input phis (10,12) to float64.

    ╭─────►Block b2
    │       10: φᵀ r0 (n4, n29) (compressed) → (x), 3 uses
    │       12: φᵀ r2 (n6, n39) (compressed) → (x), 6 uses
        ...
    │       13: CheckedSmiUntag [n10:(x)] → (x), 2 uses
    │       14: CheckedSmiUntag [n12:(x)] → (x), 1 uses
        ...
    │╭──────17: BranchIfToBooleanTrue [n16:(x)] b3 b9
        ...
    ││ │    29: φᵀ <accumulator> (n10, n12) (compressed) → (x), 4 uses
        ...
    ││ │    33: UnsafeSmiUntag [n29:(x)] → (x), 1 uses

Doing so could invalidate the `UnsafeSmiUntag` instruction.

This can only happen when hoisting the untagging out of the loop, as
this will remove the original `CheckedSmiUntag` instruction.

Fixed: 348567825
Change-Id: I2d7f2c3b544f991be9850d44bf11c3f632a0bb46