[deoptimizer][wasm] Fix deopt loop in case of non-inlineable targets
Imported functions and functions from other modules/instances are not
inlineable. Previously, the TransitiveFeedbackProcessor would just
ignore such targets.
However, with deopts, in these cases we need to make sure to switch to
the old behavior of generating a "regular" call_ref / call_indirect as
the slow path (not a deopt node) to prevent repeated deopts.

Bug: 42204618
Change-Id: I4b899b10d428b7d7507f10e992a07105eba0934e