[heap] Parallelize trivial clearing of weakrefs
Clearing weak references accounts for a large portion of GC's atomic
pause. A large fraction of this time is spent clearing weak references
that were recorded during marking, after we re-check that the referenced
objects have not been marked.

With this CL, weak references that are candidates for clearing are
distinguished in two categories: trivial and non-trivial. Trivial
weak references can be cleared in parallel, thus shortening the total
length of the atomic pause. Non-trivial weak references (that cannot be
cleared in parallel) are maps whose clearing involves custom weakness
logic.

Bug: 348181803
Change-Id: I72f338b48dc9691f77d2adeb7658dc6ab2989e71