Merged: [wasm] Fix invalid use of the generic wasm-to-js wrapper
When creating a WasmFuncRef, we assumed that the callee was a regular JS
function and used the generic wasm-to-js wrapper if the signature
matched. However some import call kinds are not supported by the generic
wrapper such as constructors. Recompute the import call kind and use the
optimized wasm-to-js wrapper if needed.

R=ahaas@chromium.org

Fixed: 348963663
(cherry picked from commit a055b948ccf12462a9bedda2103157342d6b2627)

Change-Id: I780314c8f9b8ee1789b657c24cbb96900a6f7aac