[heap] Parallelize trivial clearing of maps
This CL allows most weak references pointing to maps to be cleared in
parallel, during the trivial clearing phase. Non-trivial maps are those
contained in Map, TransitionArray or DescriptorArray host objects.

WARNING: The triviality check is expensive and it is doubtful that this
change improves performance. Let's not land this now. !!!

Bug: 348181803
Change-Id: I13776398461b1fd56933d0f554c27d6898e38505