[regexp] Create atom for lead surrogate.
When desugaring surrogate pairs we often end up with
a sequence that is a single lead surrogate followed by
a char class for the trail surrogate. In this case we
should generate an atom (literal char) instead of a tiny
char class for the lead surrogate, which causes the
TextNode to emit code that tests the simple lead
surrogate before testing for the more complicated trail
surrogate char class.

Change-Id: I42638f25c7e025a787e9a98e8c775efe4e3d9aad