[d8] Align d8 console.time* behavior with specification and Chrome.
In d8 we didn't implement `console.timeLog()` and
`console.timeStamp()` was behaving somewhat like `console.timeLog()`
(but was always using the default timer). This CL aligns the behavior
in d8 with the behavior we have in browsers and Node.js to remove
surprises and make testing easier.

This is basically a follow-up polish to https://crrev.com/c/702459
which added the `console.time*` functions in d8 initially.

Bug: 350443026
Change-Id: I4a658e947bed061afdcbf4f149b17a90b85cbdb9