[console] `console.timeStamp()` shouldn't trigger `LogEventCallback`.
Following up on https://crrev.com/c/5671342 this CL also fixes the
misuse of `console.timeStamp()` for the `LogEventCallback`. The magic
trio here is `time()`, `timeLog()`, and `timeEnd()`, and this CL fixes
it accordingly.

The `console.timeStamp()` API should only be concerned with the
browser's Performance tool, as per documentation[^1].

[^1]: http://go/mdn/API/console/timestamp_static

Bug: 350443026
Change-Id: I7862e871080cdd161be9e2628fbddb256d347866