[heap] Parallelize trivial clearing of maps
This CL allows most weak references pointing to maps to be cleared in
parallel, during the trivial clearing phase. Non-trivial maps are
those contained in Map, TransitionArray or DescriptorArray host
objects.

The triviality check does not come for free, so it is doubtful that
this change improves performance per se. But it is important for
further parallelization in a subsequent CL.

Bug: 348181803
Bug: 351469551
Change-Id: I13776398461b1fd56933d0f554c27d6898e38505