[turboshaft] Support RestLength in FrameStates
This CL adds infrastructure to support RestLength in FrameStates. It
doesn't insert any uses of it, but this will be useful when using
Maglev as a frontend and relying on Maglev's escape analysis.

Additionally, we might want to consider updating Turbofan's escape
analysis to take advantage of this as well.

Bug: chromium:42204525
Change-Id: I5da41126603a8a7cb9d53156adac94bdb761a731