Introduce ScriptCompiler::StartConsumingCodeCacheOnBackground()
There is DCHECK_NO_SCRIPT_NO_EXCEPTION() in the
ScriptCompiler::StartConsumingCodeCache() method. So we can't use this
method to start the consuming code cache task from a background thread.
So this CL introduces a new StartConsumingCodeCacheOnBackground() method
which doesn't have any DCHECK_NO_SCRIPT_NO_EXCEPTION() in it.

This method will be used in the Chromium side CL:
https://chromium-review.googlesource.com/c/chromium/src/+/5373700

Bug: 40244488

Change-Id: Ib825b5d162e7fe261505128c33c3f849be65ea00