[turbofan] Wire up String constructor exceptions correctly
When reducing the String constructor into ToString and
CreateStringWrapper, we need to make sure that the ToString's exception
edges (control and effect) are correctly connected to the String
constructor's exception edges. This is already done a few times in the
JSCallReducer (by introducing new IfSuccess/IfException and merging them
in, rather than by trying to reuse the existing nodes), so follow the
same pattern here.

Fixed: 352407772
Change-Id: Ibb8925da991f7439de6c95f628908d96c2388292