[compile hints] Add CompileOptions for compile hints magic
The embedder can direct via CompileOptions whether the eager
compilation magic comment should be obeyed. This allows the
embedder to selectively obey the magic comment, e.g., only
for compilations when we will later generate a code cache.

In addition, change the flag name to the one used in the
explainer:
https://github.com/explainers-by-googlers/explicit-javascript-compile-hints-file-based .

Bug: 42203853
Change-Id: Iaad3438b5e3b9a21ed42d02bd71374a0f7189cb2