[console] Make `console.timeEnd()` performance consistent.
As noted in https://crbug.com/41433391 and illustrated by the sample
page attached to that bug, the performance of `console.timeEnd()` can
vary greatly depending on the depth of the call stack leading to its
invocation. This can skew the results when using `console.timeEnd()` to
measure performance of certain parts of a page.

With this CL we now ensure that `console.timeEnd()` generates a console
message that always contains exactly one call frame (the top most frame).

The CL also changes the stack traces for a few other APIs, ensuring that
we don't collect any stack trace for `console.clear()`, and that we
always (even when the debugger is not attached) collect a full stack for
`console.trace()`.

Fixed: 41433391
Change-Id: I8e8076682e0c3eafcdfd1b8bc2ac184cfd980bae